+++ nodeconductor/cost_tracking/__init__.py
@@ -1,26 +1,8 @@
 """
 Cost tracking - add-on for NC plugins.
+Allows to calculate price estimates for resources from your plugin.
 
-
-Add-on adds next functional to plugin:
-
- - calculate and store price estimate for each resource, service, project, customer.
- - register resource consumables (ex: CPU, storage for VMs) prices and show
-   their prices for resource cost calculation.
- - get resource used items for current moment.
-
-
-Add-on connection:
-
-    1. Implement CostTrackingBackend interface.
-
-        class OpenStackCostTrackingBackend(CostTrackingBackend):
-            ...
-
-    2. Add application to add-on register.
-
-        CostTrackingRegister.register(self.label, cost_tracking.OpenStackCostTrackingBackend)
-
+Check developer guide for more details.
 """
 import logging
 from decimal import Decimal
@@ -34,6 +16,29 @@ default_app_config = 'nodeconductor.cost_tracking.apps.CostTrackingConfig'
 logger = logging.getLogger(__name__)
 
 
+class ConsumableItem(object):
+
+    def __init__(self, item_type, key, name=None, units='', default_price=0):
+        self.item_type = item_type
+        self.key = key
+        self.default_price = default_price
+        self.name = name if name is not None else '%s: %s' % (item_type, key)
+        self.units = units
+
+    def __str__(self):
+        return self.name
+
+    def __hash__(self):
+        return hash((self.item_type, self.key))
+
+    def __eq__(self, other):
+        return (self.item_type, self.key) == (other.item_type, other.key)
+
+    def __ne__(self, other):
+        # Not strictly necessary, but to avoid having both x==y and x!=y True at the same time
+        return not(self == other)
+
+
 class CostTrackingStrategy(object):
     """ Describes all methods that should be implemented to enable cost
         tracking for particular resource.
@@ -41,18 +46,62 @@ class CostTrackingStrategy(object):
     resource_class = NotImplemented
 
     @classmethod
-    def get_consumables(cls, resource):
+    def get_configuration(cls, resource):
         """ Return dictionary of consumables that are used by resource.
 
-            Example: {'storage': 10240, 'ram': 1024, ...}
+            Dictionary key - ConsumableItem instance.
+            Dictionary value - how many units of consumable is used by resource.
+            Example: {
+                ConsumableItem('storage', '1 MB'): 10240,
+                ConsumableItem('flavor', 'small'): 1,
+                ...
+            }
         """
         return {}
+
+    @classmethod
+    def get_consumable_items(cls):
+        """ Return list of all possible consumed items.
+
+            Output format:
+            [
+                ConsumableItem(
+                    item_type=<type of consumable>,
+                    key=<consumable name>,
+                    name=<item pretty name, that will be visible for user>,
+                    units=<consumable units (MB, GB, points, etc.>,
+                    default_price=<price for consumable usage per hour>,
+                )
+                ...
+            ]
+            Output example:
+            [
+                ConsumableItem(
+                    item_type="storage"
+                    key="1 MB",
+                    units="MB",
+                    name="1 MB of storage",
+                    default_price=0.5,
+                ),
+                ConsumableItem(
+                    item_type="flavor"
+                    key="small",
+                    name="Small flavor",
+                ),
+                ...
+            ]
+        """
+        return []
+
+
+class ResourceNotRegisteredError(TypeError):
+    pass
 
 
 class CostTrackingRegister(object):
     """ Register of all connected NC plugins """
 
-    _register = {}
+    _register = {}  # deprecated
     registered_resources = {}
 
     @classmethod
@@ -60,13 +109,32 @@ class CostTrackingRegister(object):
         cls.registered_resources[strategy.resource_class] = strategy
 
     @classmethod
-    def get_consumables(cls, resource):
+    def _get_strategy(cls, resource_class):
         try:
-            strategy = cls.registered_resources[resource.__class__]
+            return cls.registered_resources[resource_class]
         except KeyError:
-            raise TypeError('Resource %s is not registered for cost tracking. Make sure that its strategy '
-                            'is added to CostTrackingRegister' % resource.__class__.__name__)
-        return strategy.get_consumables(resource)
+            raise ResourceNotRegisteredError('Resource %s is not registered for cost tracking. Make sure that its '
+                                             'strategy is added to CostTrackingRegister' % resource_class.__name__)
+
+    @classmethod
+    def get_configuration(cls, resource):
+        """ Return how much consumables are used by resource with current configuration.
+
+            Output example:
+            {
+                <ConsumableItem instance>: <usage>,
+                <ConsumableItem instance>: <usage>,
+                ...
+            }
+        """
+        strategy = cls._get_strategy(resource.__class__)
+        return strategy.get_configuration(resource)
+
+    @classmethod
+    def get_consumable_items(cls, resource_class):
+        """ Get all possible consumable items for given resource class """
+        strategy = cls._get_strategy(resource_class)
+        return strategy.get_consumable_items()
 
     # XXX: deprecated. Should be removed.
     @classmethod